Moved libpatterns into unibuild
