+++ unibuild/libpatterns.py
